move rdm_send_response()
